Search Query Params to URL and other stuff
